+++ oc-includes/osclass/classes/controller/CWebLogin.php
@@ -230,7 +230,7 @@ class CWebLogin extends BaseModel
                 $this->redirectTo(osc_user_login_url());
                 break;
             case ('recover'):        //form to recover the password (in this case we have the form in /gui/)
-                $this->doView('user-recover.php');
+                $this->doView(osc_locate_template(array('user-recover.php'), 'user-recover'));
                 break;
             case ('recover_post'):   //post execution to recover the password
                 osc_csrf_check();
@@ -284,7 +284,7 @@ class CWebLogin extends BaseModel
                 $user = User::newInstance()
                     ->findByIdPasswordSecret(Params::getParam('userId'), Params::getParam('code'));
                 if ($user) {
-                    $this->doView('user-forgot_password.php');
+                    $this->doView(osc_locate_template(array('user-forgot_password.php'), 'user-forgot_password'));
                 } else {
                     osc_add_flash_error_message(_m('Sorry, the link is not valid'));
                     $this->redirectTo(osc_base_url());
@@ -346,7 +346,7 @@ class CWebLogin extends BaseModel
                 if (osc_logged_user_id()) {
                     $this->redirectTo(osc_user_dashboard_url());
                 }
-                $this->doView('user-login.php');
+                $this->doView(osc_locate_template(array('user-login.php'), 'user-login'));
         }
     }